+++ core/define/src/mill/define/EvaluatorProxy.scala
@@ -2,6 +2,7 @@ package mill.define
 
 import mill.api.*
 import mill.define.*
+import mill.runner.api.*
 final class EvaluatorProxy(delegate: => Evaluator) extends Evaluator {
   override def allowPositionalCommandArgs = delegate.allowPositionalCommandArgs
   override def selectiveExecution = delegate.selectiveExecution
