+++ runner/api/src/BspContextApi.scala
@@ -1,0 +1,7 @@
+
+package mill.runner.api
+
+
+private[mill] object BspContextApi {
+  @volatile var bspServerHandle: BspServerHandle = null
+}
