+++ runner/api/src/SemanticDbJavaModuleApi.scala
@@ -1,0 +1,20 @@
+package mill.runner.api
+object SemanticDbJavaModuleApi {
+  val buildTimeJavaSemanticDbVersion = Versions.semanticDbJavaVersion
+  val buildTimeSemanticDbVersion = Versions.semanticDBVersion
+
+  private[mill] val contextSemanticDbVersion: InheritableThreadLocal[Option[String]] =
+    new InheritableThreadLocal[Option[String]] {
+      protected override def initialValue(): Option[String] = None.asInstanceOf[Option[String]]
+    }
+
+  private[mill] val contextJavaSemanticDbVersion: InheritableThreadLocal[Option[String]] =
+    new InheritableThreadLocal[Option[String]] {
+      protected override def initialValue(): Option[String] = None.asInstanceOf[Option[String]]
+    }
+
+  private[mill] def resetContext(): Unit = {
+    contextJavaSemanticDbVersion.set(None)
+    contextSemanticDbVersion.set(None)
+  }
+}
