+++ runner/package.mill
@@ -10,8 +10,25 @@ import mill.contrib.buildinfo.BuildInfo
  * Mostly tested by [[build.integration]] and [[build.example]] tests.
  */
 object `package` extends RootModule with build.MillPublishScalaModule with BuildInfo {
-  object api extends build.MillPublishScalaModule {
+  object api extends build.MillPublishScalaModule with BuildInfo {
     def moduleDeps = Seq(build.core.constants)
+    def ivyDeps = Seq(build.Deps.sbtTestInterface)
+
+    def buildInfoObjectName = "Versions"
+
+    def buildInfoPackageName = "mill.runner.api"
+    def buildInfoMembers = Seq(
+      BuildInfo.Value(
+        "semanticDBVersion",
+        build.Deps.semanticDBscala.version,
+        "SemanticDB version."
+      ),
+      BuildInfo.Value(
+        "semanticDbJavaVersion",
+        build.Deps.semanticDbJava.version,
+        "Java SemanticDB plugin version."
+      ),
+    )
   }
 
   object `worker-api` extends build.MillPublishScalaModule {
