+++ scalalib/src/mill/scalalib/SemanticDbJavaModule.scala
@@ -1,6 +1,7 @@
 package mill.scalalib
 
 import mill.api.{PathRef, Result, experimental}
+import mill.runner.api.SemanticDbJavaModuleApi
 import mill.define.ModuleRef
 import mill.util.BuildInfo
 import mill.scalalib.api.{CompilationResult, Versions, JvmWorkerUtil}
@@ -24,19 +25,19 @@ trait SemanticDbJavaModule extends CoursierModule {
   def compileClasspath: T[Seq[PathRef]]
 
   def semanticDbVersion: T[String] = Task.Input {
-    val builtin = SemanticDbJavaModule.buildTimeSemanticDbVersion
+    val builtin = SemanticDbJavaModuleApi.buildTimeSemanticDbVersion
     val requested = Task.env.getOrElse[String](
       "SEMANTICDB_VERSION",
-      SemanticDbJavaModule.contextSemanticDbVersion.get().getOrElse(builtin)
+      SemanticDbJavaModuleApi.contextSemanticDbVersion.get().getOrElse(builtin)
     )
     Version.chooseNewest(requested, builtin)(Version.IgnoreQualifierOrdering)
   }
 
   def semanticDbJavaVersion: T[String] = Task.Input {
-    val builtin = SemanticDbJavaModule.buildTimeJavaSemanticDbVersion
+    val builtin = SemanticDbJavaModuleApi.buildTimeJavaSemanticDbVersion
     val requested = Task.env.getOrElse[String](
       "JAVASEMANTICDB_VERSION",
-      SemanticDbJavaModule.contextJavaSemanticDbVersion.get().getOrElse(builtin)
+      SemanticDbJavaModuleApi.contextJavaSemanticDbVersion.get().getOrElse(builtin)
     )
     Version.chooseNewest(requested, builtin)(Version.IgnoreQualifierOrdering)
   }
@@ -161,23 +162,6 @@ trait SemanticDbJavaModule extends CoursierModule {
 }
 
 object SemanticDbJavaModule {
-  val buildTimeJavaSemanticDbVersion = Versions.semanticDbJavaVersion
-  val buildTimeSemanticDbVersion = Versions.semanticDBVersion
-
-  private[mill] val contextSemanticDbVersion: InheritableThreadLocal[Option[String]] =
-    new InheritableThreadLocal[Option[String]] {
-      protected override def initialValue(): Option[String] = None.asInstanceOf[Option[String]]
-    }
-
-  private[mill] val contextJavaSemanticDbVersion: InheritableThreadLocal[Option[String]] =
-    new InheritableThreadLocal[Option[String]] {
-      protected override def initialValue(): Option[String] = None.asInstanceOf[Option[String]]
-    }
-
-  private[mill] def resetContext(): Unit = {
-    contextJavaSemanticDbVersion.set(None)
-    contextSemanticDbVersion.set(None)
-  }
 
   def javacOptionsTask(javacOptions: Seq[String], semanticDbJavaVersion: String)(implicit
       ctx: mill.api.Ctx
