+++ testrunner/src/mill/testrunner/TestRunnerMain0.scala
@@ -1,6 +1,7 @@
 package mill.testrunner
 
-import mill.api.{DummyTestReporter, internal}
+import mill.api.internal
+import mill.runner.api.DummyTestReporter
 
 @internal object TestRunnerMain0 {
   def main0(args: Array[String], classLoader: ClassLoader): Unit = {
